requirements.txt
asgiref==3.6.0
build==0.10.0
click==8.1.3
colorama==0.4.6
dj-database-url==1.3.0
Django==4.1.7
fire==0.5.0
markdown-it-py==2.2.0
mdurl==0.1.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
Pygments==2.14.0
pyproject_hooks==1.0.0
python-decouple==3.8
PyYAML==6.0
rich==13.3.3
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
typing_extensions==4.5.0
tzdata==2023.2
